more code refactor, need to work on handlebar views next

+++ public/javascript/delete-post.js
@@ -1,18 +1,27 @@
+const { DELETE } = require("sequelize/types/lib/query-types");
+
 async function deleteFormHandler(event) {
   event.preventDefault();
 
-  const id = window.location.toString().split('/')[
-    window.location.toString().split('/').length - 1
+  const id = window.location.toString().split("/")[
+    window.location.toString().split("/").length - 1
   ];
-  const response = await fetch(`/api/posts/${id}`, {
-    method: 'DELETE'
-  });
 
-  if (response.ok) {
-    document.location.replace('/dashboard/');
-  } else {
-    alert(response.statusText);
+  const verifyDelete = confirm("Delete?");
+
+  if (verifyDelete) {
+    const response = await fetch(`/api/posts${id}`, {
+      method: "DELETE",
+    });
+
+    if (response.ok) {
+      document.location.replace("/dashboard/");
+    } else {
+      alert(response.statusText);
+    }
   }
 }
 
-document.querySelector('.delete-post-btn').addEventListener('click', deleteFormHandler);
+document
+  .querySelector(".delete-post-btn")
+  .addEventListener("click", deleteFormHandler);
